Add option to include pending txn count in nonce
- `CAPE_RELAYER_NONCE_COUNT_RULE` can be set to `mined` or `pending`.
- Add CLI argument for `CAPE_RELAYER_PORT`.
- Change some port variables to u16 type.
- Run `test_relay` with both nonce count rules.

+++ wallet/src/wallet-api/main.rs
@@ -171,7 +171,7 @@ mod tests {
             &self.options
         }
 
-        async fn wait(port: u64) {
+        async fn wait(port: u16) {
             retry(|| async move {
                 // Use a one-off request, rather than going through the client, because we want to
                 // skip the middleware, which can cause connect() to return an Err() even if the
